requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf_spectacular==0.26.5
Requests==2.31.0
